+++ crates/sdmx_json/src/macros/item_scheme.rs
@@ -3,23 +3,18 @@ macro_rules! impl_item_scheme {
 		$(
 			impl $crate::structure::ItemScheme for $struct {
 				type Item = $item;
-
 				fn is_partial(&self) -> Option<bool> {
 					self.is_partial
 				}
-
 				fn items(&self) -> Option<&Vec<Self::Item>> {
 					self.$field.as_ref()
 				}
-
 				fn set_items(&mut self, items: Option<Vec<Self::Item>>) {
 					self.$field = items;
 				}
-
 				fn clear_items(&mut self) {
 					self.$field = None;
 				}
-
 				fn contains_items(&self) -> bool {
 					match &self.$field {
 						Some(items) => !items.is_empty(),